requirements.txt
Flask==2.1.1
Pillow==9.0.0
matplotlib>=2.1.0
pandas>= 0.22.0
scikit-learn>=0.19.1
scipy>=1.0.0
svgwrite>=1.1.12

